requirements.txt
pyTelegramBotAPI==4.15.2
google-generativeai==0.3.2
flask==2.3.3
requests==2.31.0